Extract Swarm logic and allow full reset

+++ app/index.js
@@ -1,104 +1,15 @@
-const { app, BrowserWindow, Menu, ipcMain } = require('electron')
+const { app, BrowserWindow, dialog, ipcMain, Menu } = require('electron')
 const { appReady, is } = require('electron-util')
 const Store = require('electron-store')
-const execa = require('execa')
-const { mkdirSync, writeFileSync, readFileSync, readdirSync, existsSync, createWriteStream } = require('fs')
 const getPort = require('get-port')
 const createOnyxServer = require('onyx-server').default
 const StaticServer = require('static-server')
 const url = require('url')
 const path = require('path')
-const os = require('os')
-
-// Note: always use `path.join()` to make sure delimiters work cross-platform
-// Some platform-specific logic may be needed here, ex using geth.exe on Windows,
-// for this you can use electron-util: https://github.com/sindresorhus/electron-util/blob/master/index.js#L17-L19
-const dataDir = path.join(app.getPath('userData'), 'data')
-const pwdPath = path.join(dataDir, 'pwd')
-const keystorePath = path.join(dataDir, 'keystore')
-const logPath = path.join(dataDir, 'node.log')
-
-const getBinPath = (name) => {
-  const platform = {
-    darwin: 'mac',
-    linux: 'linux',
-    win32: 'win',
-  }[os.platform()]
-  return is.development
-  ? path.join(app.getAppPath(), '..', 'bin', `${name}-${platform}`)
-  : path.join(process.resourcesPath, 'bin', name)
-}
-
-const setupGeth = async () => {
-  console.log("setupGeth() called")
-  const gethPath = getBinPath('geth')
-  try {
-    mkdirSync(dataDir)
-  } catch (err) {
-    // Likely directory already exists, better check though
-    // See https://nodejs.org/api/fs.html#fs_fs_mkdir_path_mode_callback
-  }
-
-  // Could be made async, shouldn't be an issue though
-  writeFileSync(pwdPath, 'secret')
-
-  // Create the account - other commands should work the same way
-  const res = await execa(gethPath, [
-    '--datadir', dataDir,
-    '--password', pwdPath,
-    'account',  'new',
-  ])
-  console.log(res)
-}
-
-const setupSwarm = async () => {
-  console.log("setupSwarm() called")
-  const swarmPath = getBinPath('swarm')
-  const keystoreFiles = readdirSync(keystorePath)
-  const keyFileName = keystoreFiles[0]
-  const keyFilePath = path.join(keystorePath, keyFileName)
-  const keystore = JSON.parse(readFileSync(keyFilePath, 'utf8'))
-
-  return await new Promise((resolve, reject) => {
-    const proc = execa(swarmPath, [
-      '--datadir', dataDir,
-      '--password', pwdPath,
-      '--bzzaccount', keystore.address,
-      '--pss',
-      '--verbosity', '4',
-      '--bzznetworkid', '922',
-      '--bootnodes', 'enode://e834e83b4ed693b98d1a31d47b54f75043734c6c77d81137830e657e8b005a8f13b4833efddbd534f2c06636574d1305773648f1f39dd16c5145d18402c6bca3@52.51.239.180:30399',
-      '--ws',
-      '--wsorigins', '*',
-    ])
-
-    proc.once('error', (error) => {
-      console.log('Failed to start node', error)
-      reject(error)
-    })
-
-    proc.stderr.pipe(createWriteStream(logPath, { flags: 'a' }))
-
-    proc.stdout.on('data', (data) => {
-      const dataStr = data.toString().toLowerCase()
-      if (dataStr.indexOf('fatal:') >= 0) {
-        const error = new Error(`swarm error: ${data.toString()}`)
-        console.log(error)
-        reject(error)
-      }
-    })
-
-    proc.stderr.on('data', (data) => {
-      const dataStr = data.toString().toLowerCase()
-      if (dataStr.indexOf('websocket endpoint opened') >= 0) {
-        console.log('Node started')
-        resolve({ swarmProc: proc })
-      }
-    })
-  })
-}
 
 const { config } = require(path.join(__dirname, 'package.json'))
+const swarm = require('./swarm')
+
 const SWARM_WS_URL =
   process.env.SWARM_WS_URL ||
   (config && config.swarmWsUrl) ||
@@ -108,20 +19,52 @@ const SWARM_HTTP_URL =
   (config && config.swarmHttpUrl) ||
   'http://localhost:8500'
 
+let appServer, loadingWindow, mainWindow
+
+const shouldQuit = app.makeSingleInstance(() => {
+  if (mainWindow) {
+    if (mainWindow.isMinimized()) mainWindow.restore()
+    mainWindow.focus()
+  }
+})
+
+if (shouldQuit) {
+  app.quit()
+}
+
+const store = new Store({ name: is.development ? 'onyx-dev' : 'onyx' })
+
+const clearEventListeners = () => {
+  if (mainWindow) {
+    mainWindow.removeListener('ready-to-show', showFunc)
+    mainWindow.removeListener('closed', closedFunc)
+  }
+}
+
 const menu = Menu.buildFromTemplate([
   {
     label: is.macos ? 'Onyx' : 'File',
     submenu: [
       {
         label: 'Reset',
-        click: () => {
-          store.delete('wsUrl')
-          if (mainWindow != null) {
-            clearEventListeners()
-            mainWindow.close()
-            mainWindow = null
-          }
-          start()
+        click: (item, focusedWindow) => {
+          dialog.showMessageBox(
+            focusedWindow,
+            {
+              type: 'warning',
+              message:
+                'Resetting the application will clear ALL data and your local Swarm identity. Are you sure you want to continue?',
+              cancelId: 0,
+              buttons: ['Cancel', 'Confirm'],
+            },
+            async index => {
+              if (index === 1) {
+                store.delete('wsUrl')
+                await swarm.reset()
+                start()
+              }
+            },
+          )
         },
       },
       {
@@ -175,10 +118,6 @@ const menu = Menu.buildFromTemplate([
   },
 ])
 
-let appServer, loadingWindow, mainWindow, connectionError
-
-const store = new Store({ name: is.development ? 'onyx-dev' : 'onyx' })
-
 const createMainWindow = async url => {
   await appReady
 
@@ -193,6 +132,7 @@ const createMainWindow = async url => {
   }
 
   closedFunc = () => {
+    swarm.stop()
     mainWindow = null
   }
 
@@ -243,26 +183,41 @@ const start = async () => {
   const storedWsUrl = store.get('wsUrl')
   if (storedWsUrl) {
     if (storedWsUrl === 'local') {
-      // Setup a local Graphql server
+      let errorMsg
+
       try {
-        if (!existsSync(keystorePath)) {
-          await setupGeth()
-        }
+        await swarm.setup()
+      } catch (e) {
+        console.log(e.stack)
+        errorMsg =
+          'There was an issue setting up the Geth account, you may want to check if your configuration is valid'
+      }
+
+      if (!errorMsg) {
         try {
-          const { swarmProc } = await setupSwarm()
+          await swarm.start()
+        } catch (e) {
+          console.log(e.stack)
+          errorMsg =
+            'There was an issue starting local Swarm node, you may want to check if you have a Swarm node running already, or if the port 3039 is open'
+        }
+      }
+
+      if (!errorMsg) {
+        try {
           const serverPort = await startLocalOnyxServer(appPort)
           const wsUrl = `ws://localhost:${serverPort}/graphql`
           const httpUrl = `http://localhost:${serverPort}`
-          appUrl = appUrl + `/?wsUrl=${wsUrl}&httpUrl=${httpUrl}`
-        } catch(e) {
-          console.log(e.stack);
-          const errorMsg = 'There was an issue starting local Swarm node, you may want to check if you have a Swarm node running already, or if the port 3039 is open'
-          appUrl = appUrl + `/?wsUrl=${storedWsUrl}&connectionError=${errorMsg}`
+          appUrl += `/?wsUrl=${wsUrl}&httpUrl=${httpUrl}`
+        } catch (e) {
+          console.log(e.stack)
+          errorMsg =
+            'There was an issue starting local GraphQL server, you may want to check you have a Swarm node running on default port 8546, or that you specified the correct port if not using default'
         }
-      } catch (err) {
-        console.warn('err: ', err)
-        const errorMsg = 'There was an issue starting local GraphQL server, you may want to check you have a Swarm node running on default port 8546, or that you specified the correct port if not using default'
-        appUrl = appUrl + `/?wsUrl=${storedWsUrl}&connectionError=${errorMsg}`
+      }
+
+      if (errorMsg) {
+        appUrl += `/?wsUrl=${storedWsUrl}&connectionError=${errorMsg}`
         if (appServer != null) {
           appServer.stop()
         }
@@ -270,18 +225,14 @@ const start = async () => {
     } else {
       // Use stored remote server url
       let domain
-      if (storedWsUrl.indexOf('://') > -1) {
+      if (storedWsUrl.indexOf('://') !== -1) {
         domain = storedWsUrl.split('/')[2]
       } else if (storedWsUrl.indexOf('/') !== -1) {
         domain = storedWsUrl.split('/')[0]
       }
-      if (!domain) {
-        const errorMsg = 'Invalid ws url'
-        appUrl = appUrl + `/?wsUrl=${storedWsUrl}&connectionError=${errorMsg}`
-      } else {
-        const httpUrl = `http://${domain}`
-        appUrl = appUrl + `/?wsUrl=${storedWsUrl}&httpUrl=${httpUrl}`
-      }
+      appUrl += domain
+        ? `/?wsUrl=${storedWsUrl}&httpUrl=http://${domain}`
+        : `/?wsUrl=${storedWsUrl}&connectionError=Invalid ws url`
     }
   }
 
@@ -297,16 +248,12 @@ const start = async () => {
     }
   })
 
-  app.on('certificate-error', (event, webContents, url, error, certificate, callback) => {
-    console.warn('cert error: ', error)
-  })
-}
-
-const clearEventListeners = () => {
-  if (mainWindow) {
-    mainWindow.removeListener('ready-to-show', showFunc)
-    mainWindow.removeListener('closed', closedFunc)
-  }
+  app.on(
+    'certificate-error',
+    (event, webContents, url, error, certificate, callback) => {
+      console.warn('cert error: ', error)
+    },
+  )
 }
 
 ipcMain.on('onSetWsUrl', (e, url) => {
